scripts: Use ceiling directories in gen-version
Attempting to use contrib/pkgbuild/PKGBUILD to install syncup on Arch
Linux fails unless it is removed from the syncup tree because
scripts/gen-version gets confused about whether it's running in a Git
repository.

Make use of GIT_CEILING_DIRECTORIES to limit Git's visibility to the
current directory, which should be the root of the syncup tree.

Signed-off-by: Alex Minghella <a@minghella.net>