Find the same configuration file as mbsync(1)
As with the Notmuch configuration, it's probably best to just use the
same logic as mbsync(1) to find a configuration file. However, we also
search for an mbsync(1) configuration file in a similar way to how we
search for a notmuch(1) configuration file.

Strictly speaking, mbsync(1) already supports multiple "profiles" in a
single configuration file, but we don't handle that so it's useful to
have a similar directory layout for both programs.

Signed-off-by: Alex Minghella <a@minghella.net>